+++ packages/backend/src/index.ts
@@ -12,12 +12,12 @@ import { App, AppServices } from './app'
 import { Config } from './config/app'
 import { GraphileProducer } from './messaging/graphileProducer'
 import { createRatesService } from './rates/service'
-import { createOutgoingPaymentService } from './outgoing_payment/service'
+import { createOutgoingPaymentService } from './open_payments/payment/outgoing/service'
 import {
   createIlpPlugin,
   IlpPlugin,
   IlpPluginOptions
-} from './outgoing_payment/ilp_plugin'
+} from './open_payments/payment/outgoing/ilp_plugin'
 import { createHttpTokenService } from './httpToken/service'
 import { createAssetService } from './asset/service'
 import { createAccountingService } from './accounting/service'
